+++ yivi_core/lib/src/screens/debug/scheme_management/scheme_manager_detail_screen.dart
@@ -42,7 +42,6 @@ class SchemeManagerDetailScreen extends StatelessWidget {
           appBar: IrmaAppBar(titleString: title, hasBorder: false),
           body: YiviPinScreen(
             instruction: instruction,
-            pinBloc: EnterPinStateBloc(maxPinSize),
             maxPinSize: maxPinSize,
             onSubmit: (pin) => navigator.pop(pin),
             listener: (context, state) {