+++ yivi_core/lib/src/screens/enrollment/enrollment_screen.dart
@@ -2,12 +2,15 @@ import "package:flutter/material.dart";
 import "package:flutter/services.dart";
 import "package:flutter_bloc/flutter_bloc.dart";
 import "package:flutter_i18n/flutter_i18n.dart";
+import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:go_router/go_router.dart";
 
 import "../../data/irma_repository.dart";
 import "../../providers/irma_repository_provider.dart";
 import "../../util/navigation.dart";
 import "../../widgets/loading_indicator.dart";
+import "../pin/providers/biometric_provider.dart";
+import "../pin/widgets/biometric_opt_in_dialog.dart";
 import "accept_terms/accept_terms_screen.dart";
 import "bloc/enrollment_bloc.dart";
 import "choose_pin/choose_pin_screen.dart";
@@ -87,8 +90,28 @@ class _ProvidedEnrollmentScreen extends StatelessWidget {
             return ConfirmPinScreen(
               newPinNotifier: newPin,
               onPrevious: addOnPreviousPressed,
-              submitConfirmationPin: (pin) =>
-                  addEvent(EnrollmentPinConfirmed(pin)),
+              // After the PIN is confirmed, offer biometric unlock (once, only
+              // if a biometric is enrolled and the user hasn't already chosen).
+              // Then advance to the email step.
+              submitConfirmationPin: (pin) async {
+                final container = ProviderScope.containerOf(
+                  context,
+                  listen: false,
+                );
+                final available = await container.read(
+                  biometricAvailableProvider.future,
+                );
+                final enabled = await repo.preferences
+                    .getBiometricEnabled()
+                    .first;
+                final dismissed = await repo.preferences
+                    .getBiometricPromptDismissed()
+                    .first;
+                if (context.mounted && available && !enabled && !dismissed) {
+                  await showBiometricOptInDialog(context);
+                }
+                if (context.mounted) addEvent(EnrollmentPinConfirmed(pin));
+              },
               onPinMismatch: () {
                 showDialog(
                   barrierDismissible: false,